修正１回目

+++ app/controllers/items_controller.rb
@@ -36,8 +36,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if user_signed_in? && current_user.id == item.user_id
+      item.destroy
+      redirect_to root_path
+    else
+      redirect_to user_session_path
+    end
   end
 
   private
